refactor to src folder
