+++ frontend/www/src/games.ts
