+++ frontend/www/src/webgl/texture.ts
